+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/list/Messages.kt
@@ -29,6 +29,7 @@ import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -226,7 +227,7 @@ internal fun BoxScope.DefaultMessagesHelperContent(
 
     val coroutineScope = rememberCoroutineScope()
 
-    val firstVisibleItemIndex = lazyListState.firstVisibleItemIndex
+    val firstVisibleItemIndex = derivedStateOf { lazyListState.firstVisibleItemIndex }
 
     val focusedItemIndex = messages.indexOfFirst { it is MessageItemState && it.focusState is MessageFocused }
 
@@ -244,7 +245,7 @@ internal fun BoxScope.DefaultMessagesHelperContent(
 
         val shouldScrollToBottomOnNewMessage = shouldScrollToBottomOnNewMessage(
             focusedItemIndex,
-            firstVisibleItemIndex,
+            firstVisibleItemIndex.value,
             newMessageState,
             areNewestMessagesLoaded,
             lazyListState.isScrollInProgress
@@ -252,7 +253,7 @@ internal fun BoxScope.DefaultMessagesHelperContent(
 
         if (shouldScrollToBottomOnNewMessage) {
             coroutineScope.launch {
-                if (newMessageState == MyOwn && firstVisibleItemIndex > 5) {
+                if (newMessageState == MyOwn && firstVisibleItemIndex.value > 5) {
                     lazyListState.scrollToItem(5)
                 }
                 lazyListState.animateScrollToItem(0)
@@ -260,14 +261,14 @@ internal fun BoxScope.DefaultMessagesHelperContent(
         }
     }
 
-    if (isScrollToBottomButtonVisible(isMessageInThread, firstVisibleItemIndex, areNewestMessagesLoaded)) {
+    if (isScrollToBottomButtonVisible(isMessageInThread, firstVisibleItemIndex.value, areNewestMessagesLoaded)) {
         MessagesScrollingOption(
             unreadCount = messagesState.unreadCount,
             modifier = Modifier.align(Alignment.BottomEnd),
             onClick = {
                 scrollToBottom {
                     coroutineScope.launch {
-                        if (firstVisibleItemIndex > 5) {
+                        if (firstVisibleItemIndex.value > 5) {
                             lazyListState.scrollToItem(5)
                         }
                         lazyListState.animateScrollToItem(0)